feat: RV64I engine, 64-bit examples, PC arrow fix
- Add separate CpuState64 / engine64 (RV64I): ld/sd/lwu, addiw/addw
  family, mulhu, 6-bit shamt for slli/srli/srai, .dword/.quad directive
- Add 64-bit register display with 32/64-bit space separator
- Reorganise examples/ into 32bit/ and 64bit/ subdirectories
- Add 10 RV64 example programs that genuinely require 64-bit arithmetic
- Add tests/examples64.rs integration suite (10 tests, all passing)
- Package both example directories in GitHub Actions release archives
- Fix PC indicator: replace Unicode arrow (tofu in egui font) with a
  painted filled triangle polygon

+++ src/isa/pseudo.rs
@@ -226,7 +226,7 @@ fn expand_li(rd: Operand, imm: i32) -> Vec<RealInstr> {
     } else {
         // upper20 rounded so that sign-extended lower12 adds correctly
         let upper = ((imm as u32).wrapping_add(0x800)) >> 12;
-        let lower = imm - ((upper as i32) << 12);
+        let lower = (imm as i64 - ((upper as i64) << 12)) as i32;
         vec![
             RealInstr::new("lui", vec![rd.clone(), Operand::Imm(upper as i32)]),
             RealInstr::new("addi", vec![rd.clone(), rd, Operand::Imm(lower)]),
